requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.12
botocore==1.42.12
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
fastapi==0.124.4
filelock==3.20.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
joblib==1.5.3
kagglehub==0.3.13
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-whois==0.9.6
pytz==2025.2
pyxDamerauLevenshtein==1.9.0
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
requests-file==3.0.1
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
threadpoolctl==3.6.0
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
xgboost==3.1.2
